sito/requirements.txt
asgiref==3.5.0
Django==4.0.4
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2021.5
mysql-connector-python==8.0.29
